chore: extracted MapillaryImage and put into own file
